refactor: consolidate types and extract components
- Unify CsvPreview and AnalysisOutput types across codebase
- Extract dashboard into 3 sub-components (FileUploader, CsvPreview, AnalysisResults)
- Extract analysis pipeline to separate service module
- Create V2 features barrel export for cleaner imports
- Add JSDoc documentation to critical modules

Dashboard reduced from 877 lines to ~347 lines (~60% reduction)
